dev(preview-frontend): split page host helpers into focused modules
